mevbot/core/flashbots.py
"""Flashbots integration and bundle optimization"""
from typing import List, Dict, Optional, Union, TypedDict, Any
from web3 import Web3
from eth_account.account import Account
from eth_account.signers.local import LocalAccount
import flashbots
from flashbots import flashbot
from flashbots.types import SignTx
import numpy as np
import asyncio
import logging
from dataclasses import dataclass
import time

logger = logging.getLogger(__name__)

class SimulationResult(TypedDict):
    success: bool
    error: Optional[str]
    gas_used: int
    effective_gas_price: int
    state_changes: List[Dict[str, Any]]
    value_transfers: List[Dict[str, Any]]

@dataclass
class BundleMetrics:
    profit: float
    gas_cost: float
    success_probability: float
    builder_reputation: float
    historical_success: float

class FlashbotsManager:
    def __init__(self, 
                 w3: Web3,
                 private_key: str,
                 config: Dict):
        self.w3 = w3
        self.config = config
        self.account: LocalAccount = Account.from_key(private_key)
        
        # Initialize Flashbots provider
        if not config.get('test_mode'):
            self.flashbot = flashbot(
                self.w3,
                self.account,
                config['flashbots_relay_url']
            )
        
        # ML model for bid optimization (placeholder)
        self.bid_model = self._initialize_bid_model()
        
        # Builder reputation tracking
        self.builder_stats: Dict[str, Dict] = {}
        
        # Bundle success tracking
        self.bundle_history: List[Dict] = []
        
        # Performance optimization
        self.enable_parallel_simulation = True
        self.max_parallel_sims = 4
        
    def _initialize_bid_model(self):
        """Initialize ML model for bid optimization"""
        return None

    def _calculate_profit_per_gas(self, bundle: List[Dict]) -> float:
        """Calculate expected profit per gas unit"""
        try:
            # In test mode, use a simple calculation
            if self.config.get('test_mode'):
                total_value = sum(tx.get('value', 0) for tx in bundle)
                total_gas = sum(tx.get('gas', 21000) for tx in bundle)
                return (total_value * 0.001) / total_gas if total_gas > 0 else 0
                
            # Calculate total value transferred
            total_value = 0
            total_gas = 0
            
            for tx in bundle:
                # Add transaction value
                total_value += tx.get('value', 0)
                
                # Add gas cost
                gas = tx.get('gas', 21000)
                gas_price = tx.get('maxFeePerGas', tx.get('gasPrice', 0))
                total_gas += gas
                
                # Subtract gas cost from profit
                total_value -= gas * gas_price
            
            # Calculate profit per gas
            return total_value / total_gas if total_gas > 0 else 0
            
        except Exception as e:
            logger.error(f"Error calculating profit per gas: {e}")
            return 0

    async def _optimize_bundle(self, transactions: List[Dict]) -> Optional[List[Dict]]:
        """Optimize transaction bundle for maximum profit"""
        try:
            if not transactions:
                return None
                
            # Calculate profit per gas for the bundle
            profit_per_gas = self._calculate_profit_per_gas(transactions)
            
            if profit_per_gas <= 0:
                logger.warning("Bundle not profitable")
                return None
                
            # In test mode, return a simple optimized bundle
            if self.config.get('test_mode'):
                return transactions
                
            # Sort transactions by profit per gas
            sorted_txs = sorted(
                transactions,
                key=lambda tx: self._calculate_profit_per_gas([tx]),
                reverse=True
            )
            
            # Optimize gas prices
            optimized_txs = []
            base_fee = await self.w3.eth.get_block('latest').base_fee_per_gas
            
            for tx in sorted_txs:
                # Calculate optimal gas price
                optimal_gas = max(
                    int(base_fee * 1.2),  # 20% above base fee
                    tx.get('maxFeePerGas', tx.get('gasPrice', 0))
                )
                
                # Update transaction
                optimized_tx = dict(tx)
                optimized_tx['maxFeePerGas'] = optimal_gas
                optimized_tx['maxPriorityFeePerGas'] = int(optimal_gas * 0.1)
                
                optimized_txs.append(optimized_tx)
            
            return optimized_txs
            
        except Exception as e:
            logger.error(f"Bundle optimization failed: {e}")
            return None

    async def submit_bundle(self, 
                          transactions: List[Dict],
                          target_block: int) -> Optional[str]:
        """
        Submit bundle to Flashbots with optimized bidding
        """
        try:
            # Optimize bundle
            optimized_bundle = await self._optimize_bundle(transactions)
            if not optimized_bundle:
                logger.warning("Bundle optimization failed")
                return None
            
            # Simulate bundle
            simulation = await self._simulate_bundle(optimized_bundle, target_block)
            if not simulation or not simulation.success:
                logger.warning("Bundle simulation failed")
                return None
            
            # Calculate optimal bid
            optimal_bid = self._calculate_optimal_bid(
                simulation,
                optimized_bundle,
                target_block
            )
            
            # Prepare bundle with bid
            signed_bundle = await self._prepare_bundle(
                optimized_bundle,
                optimal_bid,
                target_block
            )
            
            # Submit to Flashbots
            result = await self.flashbot.send_bundle(
                signed_bundle,
                target_block_number=target_block
            )
            
            # Track submission
            self._track_bundle_submission(result, signed_bundle, target_block)
            
            return result.bundle_hash
            
        except Exception as e:
            logger.error(f"Error submitting bundle: {e}")
            return None
    
    async def _parallel_bundle_optimization(self, 
                                         transactions: List[Dict]) -> Optional[List[Dict]]:
        """
        Optimize bundle composition using parallel simulations
        """
        best_bundle = None
        best_metrics = None
        
        # Generate bundle candidates
        candidates = self._generate_bundle_candidates(transactions)
        
        # Simulate candidates in parallel
        async with asyncio.Semaphore(self.max_parallel_sims) as sem:
            tasks = []
            for candidate in candidates:
                task = asyncio.create_task(
                    self._evaluate_bundle_candidate(candidate, sem)
                )
                tasks.append(task)
            
            results = await asyncio.gather(*tasks)
            
            # Select best bundle
            for bundle, metrics in results:
                if metrics and (not best_metrics or 
                              metrics.profit > best_metrics.profit):
                    best_bundle = bundle
                    best_metrics = metrics
        
        return best_bundle
    
    def _calculate_optimal_bid(self,
                             simulation: SimulationResult,
                             bundle: List[Dict],
                             target_block: int) -> int:
        """
        Calculate optimal bid price using ML model
        """
        try:
            # Features for bid calculation
            features = np.array([
                simulation.total_gas_used,
                self._get_current_base_fee(),
                self._get_network_congestion(),
                self._calculate_bundle_value(bundle),
                self._get_historical_success_rate(target_block),
                self._get_builder_reputation_score()
            ])
            
            if self.bid_model:
                # Use ML model for prediction
                optimal_bid = self.bid_model.predict(features.reshape(1, -1))[0]
            else:
                # Fallback to heuristic calculation
                optimal_bid = self._calculate_heuristic_bid(features)
            
            return int(optimal_bid)
            
        except Exception as e:
            logger.error(f"Error calculating optimal bid: {e}")
            return self._get_default_bid()
    
    async def _simulate_bundle(self,
                             bundle: List[Dict],
                             target_block: int) -> Optional[SimulationResult]:
        """
        Simulate bundle execution
        """
        try:
            state_block_tag = target_block - 1
            result = await self.flashbot.simulate(
                bundle,
                block_tag=state_block_tag
            )
            
            return result
            
        except Exception as e:
            logger.error(f"Bundle simulation failed: {e}")
            return None
    
    def _track_bundle_submission(self,
                               result: Dict,
                               bundle: List[Dict],
                               target_block: int):
        """
        Track bundle submission for analysis
        """
        submission_data = {
            'timestamp': time.time(),
            'target_block': target_block,
            'bundle_hash': result.bundle_hash,
            'simulation_success': result.simulation_success,
            'simulation_error': result.simulation_error,
            'bundle_size': len(bundle)
        }
        
        self.bundle_history.append(submission_data)
        
        # Cleanup old history
        if len(self.bundle_history) > 1000:
            self.bundle_history = self.bundle_history[-1000:]
    
    def _get_builder_reputation_score(self) -> float:
        """
        Calculate builder reputation score
        """
        if not self.builder_stats:
            return 0.5
        
        # Calculate weighted average of builder success rates
        total_weight = 0
        weighted_score = 0
        
        for builder, stats in self.builder_stats.items():
            weight = stats.get('blocks_built', 0)
            score = stats.get('success_rate', 0.5)
            
            weighted_score += weight * score
            total_weight += weight
        
        return weighted_score / total_weight if total_weight > 0 else 0.5
    
    def _get_historical_success_rate(self, target_block: int) -> float:
        """
        Get historical success rate for similar blocks
        """
        if not self.bundle_history:
            return 0.5
        
        # Calculate success rate for recent submissions
        recent_submissions = [
            b for b in self.bundle_history
            if abs(b['target_block'] - target_block) < 100
        ]
        
        if not recent_submissions:
            return 0.5
        
        success_count = sum(1 for b in recent_submissions if b['simulation_success'])
        return success_count / len(recent_submissions)
    
    def _calculate_heuristic_bid(self, features: np.ndarray) -> int:
        """
        Calculate bid using heuristic approach
        """
        base_fee = features[1]
        congestion = features[2]
        bundle_value = features[3]
        
        # Basic heuristic: bid more during congestion and for valuable bundles
        bid = base_fee * (1 + congestion) * (1 + bundle_value / 1e18)
        
        return int(bid)
